detectron2/requirements.txt
opencv_python==4.6.0.66
fastapi[all]>=0.85.1,<0.85.2
pydantic>=1.10.2,<1.10.3
uvicorn[standard]>=0.18.3,<0.18.4
python-multipart >=0.0.5,< 0.0.6
numpy
passlib>=1.7.4,<1.7.5
pydantic==1.10.2
pymongo==4.3.3
python_jose==3.3.0
websockets>=10.3,<10.4
httpx>=0.23.1,<0.23.2
bcrypt==4.0.1
